# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Event Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterInputRole;
  /** @nullable */
  studentId?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  studentId?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Event {
  id: string;
  title: string;
  description: string;
  date: string;
  location: string;
  capacity: number;
  spotsLeft: number;
  category: string;
  organizer: string;
  registeredCount: number;
  /** @nullable */
  isRegistered?: boolean | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface EventDetail {
  id: string;
  title: string;
  description: string;
  date: string;
  location: string;
  capacity: number;
  spotsLeft: number;
  category: string;
  organizer: string;
  registeredCount: number;
  /** @nullable */
  isRegistered?: boolean | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface EventInput {
  /** @minLength 3 */
  title: string;
  description: string;
  date: string;
  location: string;
  /** @minimum 1 */
  capacity: number;
  category: string;
  organizer: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface EventUpdate {
  /** @minLength 3 */
  title?: string;
  description?: string;
  date?: string;
  location?: string;
  /** @minimum 1 */
  capacity?: number;
  category?: string;
  organizer?: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Registration {
  id: string;
  eventId: string;
  userId: string;
  registeredAt: string;
}

export interface Participant {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  studentId?: string | null;
  registeredAt: string;
}

export interface AdminEvent {
  id: string;
  title: string;
  description?: string;
  date: string;
  location: string;
  capacity: number;
  spotsLeft: number;
  category: string;
  organizer: string;
  registeredCount: number;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalEvents: number;
  totalStudents: number;
  totalRegistrations: number;
  upcomingEvents: number;
  popularEvents: AdminEvent[];
}

export type ListEventsParams = {
category?: string;
search?: string;
upcoming?: string;
};