# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: College Event Management System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: events
    description: Event operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List all events
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: upcoming
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Event deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/{id}/registrations:
    post:
      operationId: registerForEvent
      tags: [events]
      summary: Register current user for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Already registered or event full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: unregisterFromEvent
      tags: [events]
      summary: Unregister current user from an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Unregistered successfully
        "404":
          description: Registration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /events/{id}/participants:
    get:
      operationId: getEventParticipants
      tags: [events]
      summary: Get participants of an event (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of participants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Participant"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/events:
    get:
      operationId: listAdminEvents
      tags: [admin]
      summary: List all events with registration counts (admin only)
      responses:
        "200":
          description: All events with registration counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminEvent"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /students/events:
    get:
      operationId: listMyRegistrations
      tags: [events]
      summary: Get events the current student is registered for
      responses:
        "200":
          description: List of registered events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, admin]
        studentId:
          type: ["string", "null"]

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
        studentId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    Event:
      type: object
      required: [id, title, description, date, location, capacity, spotsLeft, category, organizer, registeredCount, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        location:
          type: string
        capacity:
          type: integer
        spotsLeft:
          type: integer
        category:
          type: string
        organizer:
          type: string
        registeredCount:
          type: integer
        isRegistered:
          type: ["boolean", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    EventDetail:
      type: object
      required: [id, title, description, date, location, capacity, spotsLeft, category, organizer, registeredCount, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        location:
          type: string
        capacity:
          type: integer
        spotsLeft:
          type: integer
        category:
          type: string
        organizer:
          type: string
        registeredCount:
          type: integer
        isRegistered:
          type: ["boolean", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    EventInput:
      type: object
      required: [title, description, date, location, capacity, category, organizer]
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
        date:
          type: string
          format: date-time
        location:
          type: string
        capacity:
          type: integer
          minimum: 1
        category:
          type: string
        organizer:
          type: string
        imageUrl:
          type: ["string", "null"]

    EventUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 3
        description:
          type: string
        date:
          type: string
          format: date-time
        location:
          type: string
        capacity:
          type: integer
          minimum: 1
        category:
          type: string
        organizer:
          type: string
        imageUrl:
          type: ["string", "null"]

    Registration:
      type: object
      required: [id, eventId, userId, registeredAt]
      properties:
        id:
          type: string
        eventId:
          type: string
        userId:
          type: string
        registeredAt:
          type: string
          format: date-time

    Participant:
      type: object
      required: [id, name, email, registeredAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        studentId:
          type: ["string", "null"]
        registeredAt:
          type: string
          format: date-time

    AdminEvent:
      type: object
      required: [id, title, date, location, capacity, spotsLeft, category, organizer, registeredCount, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        location:
          type: string
        capacity:
          type: integer
        spotsLeft:
          type: integer
        category:
          type: string
        organizer:
          type: string
        registeredCount:
          type: integer
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    DashboardStats:
      type: object
      required: [totalEvents, totalStudents, totalRegistrations, upcomingEvents, popularEvents]
      properties:
        totalEvents:
          type: integer
        totalStudents:
          type: integer
        totalRegistrations:
          type: integer
        upcomingEvents:
          type: integer
        popularEvents:
          type: array
          items:
            $ref: "#/components/schemas/AdminEvent"