# lib/api-zod/src/generated/types/adminEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Event Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminEvent {
  id: string;
  title: string;
  description?: string;
  date: Date;
  location: string;
  capacity: number;
  spotsLeft: number;
  category: string;
  organizer: string;
  registeredCount: number;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: Date;
}