# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Event Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminEvent } from './adminEvent';

export interface DashboardStats {
  totalEvents: number;
  totalStudents: number;
  totalRegistrations: number;
  upcomingEvents: number;
  popularEvents: AdminEvent[];
}