# lib/api-zod/src/generated/types/participant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Event Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Participant {
  id: string;
  name: string;
  email: string;
  /** @nullable */
  studentId?: string | null;
  registeredAt: Date;
}