# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Event Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterInputRole } from './registerInputRole';

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterInputRole;
  /** @nullable */
  studentId?: string | null;
}