# lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * College Event Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Registration {
  id: string;
  eventId: string;
  userId: string;
  registeredAt: Date;
}